Renamed delete test; fixed pom
