+++ pom.xml
@@ -6,10 +6,12 @@
   <name>mmouse-uml-editor</name>
   <description>UML Editor created by MMouse LLC</description>
   <properties>
-     <maven.compiler.source>11</maven.compiler.source>
-     <maven.compiler.target>11</maven.compiler.target>
+    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+    <maven.compiler.source>11</maven.compiler.source>
+    <maven.compiler.target>11</maven.compiler.target>
   </properties>
   <build>
+    <sourceDirectory>src/main/java</sourceDirectory>
     <plugins>
       <plugin>
         <groupId>org.openjfx</groupId>
@@ -26,33 +28,26 @@
         <configuration>
           <source>11</source>
           <target>11</target>
-        </configuration>        
-      </plugin>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-surefire-plugin</artifactId>
-        <version>3.0.0-M3</version>
-	<configuration>
-          <forkCount>0</forkCount>
         </configuration>
       </plugin>
       <plugin>
         <groupId>org.jacoco</groupId>
         <artifactId>jacoco-maven-plugin</artifactId>
-        <version>0.7.6.201602180812</version>
+        <version>0.8.5</version>
         <executions>
           <execution>
-            <id>prepare-agent</id>
+            <id>default</id>
             <goals>
               <goal>prepare-agent</goal>
+              <goal>report</goal>
             </goals>
           </execution>
         </executions>
       </plugin>
       <plugin>
-    	<groupId>org.eluder.coveralls</groupId>
-    	<artifactId>coveralls-maven-plugin</artifactId>
-    	<version>4.3.0</version>
+        <groupId>org.eluder.coveralls</groupId>
+        <artifactId>coveralls-maven-plugin</artifactId>
+        <version>4.3.0</version>
       </plugin>
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
